spring5-recipe-app/src/test/java/com/cliff/recipeapp/services/RecipeServiceImplTest.java
package com.cliff.recipeapp.services;

import com.cliff.recipeapp.converters.RecipeCommandToRecipe;
import com.cliff.recipeapp.converters.RecipeToRecipeCommand;
import com.cliff.recipeapp.domain.Recipe;
import com.cliff.recipeapp.exceptions.NotFoundException;
import com.cliff.recipeapp.repositories.RecipeRepository;
import org.junit.Before;
import org.junit.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import java.util.HashSet;
import java.util.Optional;
import java.util.Set;

import static org.junit.Assert.*;
import static org.mockito.ArgumentMatchers.anyLong;
import static org.mockito.Mockito.*;

/**
 * @author Cliff
 * 10/7/17
 */
public class RecipeServiceImplTest {

    RecipeServiceImpl recipeService;

    @Mock
    RecipeRepository recipeRepository;

    @Mock
    RecipeToRecipeCommand recipeToRecipeCommand;

    @Mock
    RecipeCommandToRecipe recipeCommandToRecipe;


    @Before
    public void setUp() throws Exception {
        MockitoAnnotations.initMocks(this);

        recipeService = new RecipeServiceImpl(recipeRepository, recipeCommandToRecipe, recipeToRecipeCommand );
    }

    @Test
    public void getRecipeByIdTest() throws Exception {
        Recipe recipe = new Recipe();
        recipe.setId(1L);
        Optional<Recipe> recipeOptional = Optional.of(recipe);

        when(recipeRepository.findById(anyLong())).thenReturn(recipeOptional);

        Recipe recipeReturned = recipeService.findById(1L);

        assertNotNull("Null recipe returned", recipeReturned);
        verify(recipeRepository, times(1)).findById(anyLong());
        verify(recipeRepository, never()).findAll();
    }

    @Test( expected = NotFoundException.class )
    public void getRecipeByIdWhenIdNotFoundShouldThrowException() throws Exception {
        //given
        Recipe recipe = new Recipe();
        recipe.setId( 99L );
        Optional<Recipe> emptyOptional = Optional.empty();

        when( recipeRepository.findById( anyLong() )).thenReturn( emptyOptional );

        //when
        Recipe recipeReturned = recipeService.findById( 99L );
        //exception should be thrown
    }

    @Test
    public void getRecipesTest() throws Exception {

        Recipe recipe = new Recipe();
        HashSet recipesData = new HashSet();
        recipesData.add(recipe);

        when( recipeService.getAllRecipes()).thenReturn( recipesData );

        Set<Recipe> recipes = recipeService.getAllRecipes();

        assertEquals(recipes.size(), 1);
        verify(recipeRepository, times(1)).findAll();
        verify(recipeRepository, never()).findById(anyLong());
    }

    @Test
    public void testDeleteById() throws Exception {

        //given
        Long idToDelete = Long.valueOf(2L);

        //when
        recipeService.deleteById(idToDelete);

        //no 'when', since method has void return type

        //then
        verify(recipeRepository, times(1)).deleteById(anyLong());
    }

}